Fixed Writing Issue
