+++ AnisotropicKuwahara/AnisotropicKuwahara.py
@@ -1,49 +1,19 @@
 import numpy as np
 from multiprocessing import Pool
-
-class KuwaharaAnisotropic:
+from .Filter import Filter
+class KuwaharaAnisotropic(Filter):
     def __init__(self, src, structure_tensor, size=10.0, sharpness=1.0, eccentricity=0.5):
-        self.size = size
-        height, width, _ = src.shape
-        self.height = height
-        self.width = width
-        self.dst = np.zeros(shape=(height, width, 4))
+        super().__init__(src=src)
         self.sharpness = sharpness
         self.eccentricity = eccentricity
-        self.src = src
         self.structure_tensor = structure_tensor
 
     @staticmethod
     def square(x):
         return x * x
-    
-    def getDst(self):
-        return self.dst
-    
-    def doX(self, y):
-        column = np.zeros(shape=(self.width, 4))
-        try:
-            for x in range(self.width):
-                column[x] = self.process([y, x])
-        except:
-            print(f"Failure in column {str(y)}")
-        return column
-    
-    def multiProcessAll(self):
-        pool = Pool()
-        results = []
-        for y in reversed(range(self.height)):
-            print(f"Adding {str(y)} to pool")
-            results.append(pool.apply_async(self.doX, args=(y,)))
-        y = 0
-        for result in results:
-            print(result)
-            self.dst[y] = result.get()
-            print(f"Got result for {str(y)}")
-            y += 1
-
 
     def process(self, pos):
+        return np.array([0,0,0,1])
         structure_tensor = self.structure_tensor
         src = self.src
         encoded_structure_tensor = structure_tensor[pos[0], pos[1]]
